Adds two new fault tolerance tests
One for killing elected coordinator and trying to read data.
One for trying to "write" after the master server is killed.

I needed to reduce # of clusters to 4 due to this being a
sufficient number for the tests we need, and 7 clusters seems
to trigger the RamCloud C++ RPC timeout bug on 1 of 2 runs of
the integration tests.